Spread hook logic across two custom hooks and simplify TierNFT component.
